modules/TruthNews/news_analysis/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
python-dotenv>=0.19.0
